# Refactor flags mechanisms