I think It's more obvious now to understand the inheritance of flags.

Updates #395
Fix #403
Fix #404

Performance improvements:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkInheritedFlags-4     6536          5595          -14.40%
BenchmarkLocalFlags-4         3193          3235          +1.32%

benchmark                     old allocs     new allocs     delta
BenchmarkInheritedFlags-4     49             39             -20.41%
BenchmarkLocalFlags-4         23             21             -8.70%

benchmark                     old bytes     new bytes     delta
BenchmarkInheritedFlags-4     2040          1000          -50.98%
BenchmarkLocalFlags-4         1008          544           -46.03%
```

Also merge of https://github.com/spf13/pflag/pull/119 will significantly improve performance:
```
BenchmarkInheritedFlags-4   	 1000000	      1365 ns/op	       0 B/op	       0 allocs/op
BenchmarkLocalFlags-4       	 2000000	       900 ns/op	       0 B/op	       0 allocs/op
```