@bogem 2017-04-09
Also `-h, --help` is shown in usage now